move endpoint clients into providers
